requirements.txt
appdirs==1.4.4
attrs==21.2.0
black==21.6b0
certifi==2021.5.30
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
dacite==1.6.0
dataclasses-jsonschema==2.14.1
distlib==0.3.2
filelock==3.0.12
flake8==3.9.2
flake8-print==4.0.0
Flask==2.0.0
Flask-Cors==3.0.10
Flask-Script==2.0.6
identify==2.2.10
idna==2.10
isort==5.8.0
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
marshmallow==3.12.1
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
oauthlib==3.1.1
pathspec==0.8.1
pip==21.1.2
pre-commit==2.13.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==1.7.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
setuptools==57.0.0
six==1.16.0
toml==0.10.2
urllib3==1.26.5
virtualenv==20.4.7
Werkzeug==2.0.1
wheel==0.36.2
flask_swagger_ui==3.36.0
pandas==1.3.3
python-magic==0.4.24
pymongo==3.12.1
docker==5.0.3